AJout confirmatio,

+++ frontent-admin/src/app/article/components/list-article/list-article.component.ts
@@ -3,7 +3,7 @@ import {ArticleService} from "../../services/article.service";
 import { faSync } from '@fortawesome/free-solid-svg-icons';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
 import { environment } from '../../../../environments/environment';
-
+import { ConfirmationDialogService } from './../../../ui/confirmation-dialog/confirmation-dialog.service';
 @Component({
   selector: 'app-list-article',
   templateUrl: './list-article.component.html',
@@ -12,11 +12,11 @@ import { environment } from '../../../../environments/environment';
 export class ListArticleComponent implements OnInit {
   faSync = faSync;
   faTrash = faTrash;
-
+  confirmation;
   public articleList;
   public environment = environment;
 
-  constructor(private articleService: ArticleService) {
+  constructor(private articleService: ArticleService, private confirmationDialogService: ConfirmationDialogService) {
     this.articleList = [];
     this.refresh();
   }
@@ -29,8 +29,13 @@ export class ListArticleComponent implements OnInit {
   }
 
   async deleteUser(id) {
+    await this.confirmationDialogService.confirm('Please confirm..', 'Do you really want to ... ?')
+    .then(confirm => this.confirmation = confirm)
+    .catch(() => console.log('User dismissed the dialog (e.g., by using ESC, clicking the cross icon, or clicking outside the dialog)'));
     // Suppression de l'utilisateur
-    await this.articleService.deleteArticle(id);
+    if(this.confirmation){
+      await this.articleService.deleteArticle(id);
+    }
     // Puis on rafraichi la liste
     this.refresh();
   }
